`force_client_auth=false` allows client to authenticate anyway
The authentication information will be available through the `AuthContext`.
This will not impact performance if client is not provided with a certificate or if `force_client_auth=true`.

Server requests client certificate and enforces that the client presents a
certificate.

The cerificate presented by the client is verified by the gRPC framework.
(For a successful connection the client needs to present a certificate that
can be verified against the root certificate configured by the server)

The client's key certificate pair must be valid for the SSL connection to
be established.

See https://github.com/pingcap/grpc/blob/984a89c01f31c272c9b7f14f4f2f804e6d80be09/include/grpc/grpc_security_constants.h for more details