util/buf: rusty grpc_slice and its related structs
This PR wraps `grpc_slice` into a safe struct and provides convenient
and safe APIs to manipulate and transform.

This is the first step to reduce allocations in grpcio, see also #400.

This PR has a lot of common changes with #382. The most notable
difference is the implement of vtable struct.

Signed-off-by: Jay Lee <busyjaylee@gmail.com>