*: implements health check (#518)
This PR provides simple health check implementations referring to the go
version. It provides a HealthService to maintain the service statuses
and wake up watchers.

I use a standalone crate to avoid introducing protobuf specific code into the
grpcio crate. And I don't reuse grpcio-proto to avoid the dependency of
protobuf-build which just make things complicated by introducing a lot of
dependencies and resulting in a dependency circle.

Signed-off-by: Jay Lee <BusyJayLee@gmail.com>